ENH: Added MaskedParametricImageSource class and test
The MaskedParametricImageSource class is an adapter class that
presents a ParametricImageSource as one with potentially fewer
parameters. Individual parameters may be enabled or disabled.  This
capability is useful for using a ParametricImageSource in a parameter
estimation routine where some of the parameters should be treated as
fixed (disabled parameters) and the others are desired to be estimated
(enabled parameters).

Change-Id: I526f0aa491ad7ea2368cf118f8c849d57c7055df